+++ exporter.go
@@ -260,8 +260,9 @@ func main() {
 		}
 		registry.MustRegister(collector.NewNginxPlusCollector(plusClient.(*plusclient.NginxClient), "nginxplus"))
 	} else {
+		appName := fmt.Sprintf("NGINX-Prometheus-Exporter/v%v", version)
 		ossClient, err := createClientWithRetries(func() (interface{}, error) {
-			return client.NewNginxClient(httpClient, *scrapeURI)
+			return client.NewNginxClient(httpClient, *scrapeURI, appName)
 		}, *nginxRetries, nginxRetryInterval.Duration)
 		if err != nil {
 			log.Fatalf("Could not create Nginx Client: %v", err)