+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+### DEV
+
+FEATURES:
+* Set User-Agent header when calling the nginx stats endpoint.
+
 ### 0.4.2
 
 BUGFIXES: